Code tidying.
Calibration WIP.

There's an error somewhere (or maybe it's the ill-conditioned
nature of the problem, in which case, Arrrgh!) because feeding triangle corners into a scanner
with zero errors gives a sum of squared errors that is not zero.

More WIP needed.  Also needed: gradient descent function minimisation; there may well be
a Python library somewhere for this.  To be done.